+++ src/@components/MainPage/BestPiickle/index.tsx
@@ -1,9 +1,9 @@
 import { routePaths } from "../../../core/routes/path";
 import { headingTitles } from "../../../util/main/headingTitles";
+import BestPiickleCard from "../../@common/BestPiickleCard";
 import HeadingTitleContainer from "../../@common/HeadingTitleContainer";
 import useDraggingContainer from "../../@common/hooks/useDraggingContainer";
 import { useBestPiickle } from "../hooks/useBestPiickle";
-import BestPiickleCard from "./BestPiickleCard";
 import St from "./style";
 
 export default function BestPiickle() {
@@ -23,7 +23,13 @@ export default function BestPiickle() {
           {bestPiickle &&
             bestPiickle.data.slice(0, 5).map((bestPiickle, idx) => {
               return (
-                <BestPiickleCard key={bestPiickle._id} bestPiickle={bestPiickle} idx={idx} canNavigate={!isDragging} />
+                <BestPiickleCard
+                  key={bestPiickle._id}
+                  bestPiickle={bestPiickle}
+                  idx={idx}
+                  canNavigate={!isDragging}
+                  isLast={idx !== 5}
+                />
               );
             })}
         </St.SliderWrapper>
